search-service/requirements.txt
fastapi
uvicorn
supabase
openai
requests
python-dotenv
